Move TypedNode into separate module
