+++ src/decode.rs
@@ -18,8 +18,8 @@
 //! Refer to [`crate::encode`] for information on the encoding.
 
 use crate::bititer::BitIter;
-use crate::core::types::{FinalType, FinalTypeInner, TypedProgram};
-use crate::core::{Term, UntypedProgram, Value};
+use crate::core::types::{FinalType, FinalTypeInner};
+use crate::core::{Term, TypedProgram, UntypedProgram, Value};
 use crate::jet::Application;
 use crate::merkle::cmr::Cmr;
 use crate::Error;
